repos: Validate integrity of check-out against sha256sum
If a repo comes with a sha256sum field set, generate an archive from the
checked out revision and hash that using SHA-256. This allows to
identify potential supply chain attacks, e.g. when someone took over the
original repository or a used mirror and injected a modified version of
it. Various attacks are possible in that case, not only because of
weaknesses of the SHA-1 hashing mechanism that is used by default in git
and hg today.

As git and possibly also hg do not guarantee stability of archives
generated by the built-in commands, we define a custom format. It
consists of a header line for each file, listing its name, size and the
octal-encoded raw mode, terminated by a line ending. Then, either the
content of the file or the value of the symbolic link follows. The hash
is calculated over all files dumped in this format.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>