context: take minimal dependencies from `PATH`
Fixes nasty

```bash
$ kas checkout # ..
FileNotFoundError: [Errno 2] No such file or directory: 'git'
```

error when run from a nixos system or a system for which the `git`
command (or the `hg` command) is not installed system wide or is
installed from a developer environment.

Also allow us to provide `git` via a developer environment such
as `nix-env`.

Signed-off-by: Raymond Gauthier <jraygauthier@gmail.com>