repos: Never consider hex-string refspecs with SHA lengths as branches
Do not try to checkout a branch which is a hexadecimal string and has a
length that could also be a SHA-1 or SHA-256. This avoids attacks when
someone takes over a repo and injects a branch which is called like a
valid SHA.

So far, kas preferred branches over existing SHAs during checkout.
Aditya Sirish A reported this as potential attack. But even when
changing this behavior, an attacker could also drop the previously used
SHA from the repo, tricking kas to use the branch instead. The new
policy avoids this ambiguity. Users are required to use non-abbreviated
SHAs as refspec to benefit from this protection.

Note that this does not resolve the risk of attacks via collisions of
the weak SHA-1 algorithm.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>