repos: improved parsing of the 'series' file with whitespaces
When kas applies patches and the 'series' file contains an empty line or
whitespaces before comment, it fails on error FileNotFoundError, which has
nothing to do with a file operation, it's only a parsing issue.

This patch fixes issue with parsing the 'series' file. Moreover, test files
were updated to cover such case.

https://github.com/siemens/kas/issues/96

Signed-off-by: Viliam Lejcik <lejcik@gmail.com>