kas-container: detect build system on clean commands
This patch reworks the kas-container clean* logic to support passing a
kas configuration file. By that, also the default kas file
(.config.yaml) is processed in case no kas file is provided. This makes
it possible to run "kas-container clean" on an ISAR build that was
configured with "kas-container menu". Previously, "kas-container --isar
clean" needed to be executed in this situation.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: fixed shellcheck error and massaged commit message]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>