menu plugin: Add separate return button
While working with submenus, its rather inconvenient to move back to
the parent menu. The first button in the line should always be the most
commonly used one in the given context. And while this is "build" for
the main menu, "return" is most common when in a submenu.

To achieve this behavior, add a dedicated return button to submenus
rather than redefining the exit button. Furthermore, the ESC hotkey
gets context dependent semantic: exit when in the main menu, otherwise
return.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>