Add "tag" property to repo, to replace usage of refspec.
Tags usually never move, but they are mutable, so a configuration
can reference commit and tag at the same time.
Kas will check that the commit and tag match, and if not, error out.

If no commit is provided with the tag, kas will issue a warning about
the unsafeness of tags.

Signed-off-by: Félix Piédallu <felix@piedallu.me>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>