locking: generalize locking to all included files
Previously, kas only checked if a lockfile exists for the first file on
the kas commandline. This is problematic when injecting mirror
configurations by appending the site config as last file. In addition,
it makes it hard for projects with multiple configs (kas-files) to use
the locking support.

This patch removes this limitation by checking for each file for a
corresponding lock file. Hereby the same depths-first search logic is
applied as with kas includes. Each lockfile is added as it would be the
next include to process, i.e. after descending into the current
hierarchie. This is applied recursively.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>