kas(aws): add support for oauth2 worflow
This patch adds support for the OAuth 2.0 based authentication against
the AWS S3 backend. This is implemented by mapping the
AWS_WEB_IDENTITY_TOKEN_FILE and the AWS_ROLE_ARN environment variables
onto the corresponding entries in the config file. By that, also older
versions of bitbake that do not forward these variables are supported.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>