kas-container: unify error handling
This patch adds functions to report errors and warnings. These ensure,
that fatal errors terminate the program and all messages are printed to
stderr (to not interfere with KAS output on stdout). All existing
logging is ported over to these functions. By that, the inconsistent
printing to stderr and stdout is fixed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>