docs: improve documentation w.r.t empty env vars
The documentation is imprecise (if not wrong) regarding how to specify a
conditional forward of an environment variable. While stating that the
right side of the assignment needs to be None is correct in python
speak, it is NOT correct w.r.t. how none-types are defined in YAML /
JSON. Here, the the value *None* is converted to a string 'None' and
'null' specifies a string with value 'null'. Both values are different
from the nulltype constant *null*.

This patch makes the documentation more precise regarding this and also
updates the old (kas 3.2) format-changelog entry, as the implementation
has always been like that (only considering the nulltype constant).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>