fix including from transitively referenced repos
This patch fixes an issue when transitively including repositories. In
case the include references a repository (using 'repo') that is not yet
checked out, the SetupReposStep stopped as the repo is not yet in the
config dict. However, this might still be a correct specification where
simply the repo did not got downloaded yet.

The logic is now slightly changed to simply continue with the checkout
of the next repository. By that, the loop will execute until we reach a
stable state, where either all repos are checked out, or no progress was
made between two iterations (the old and new missing repos are the
same).

Reported-by: Liao Pingfang <pingfang.liao@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>