check if branch contains commit if both are set
When specifying both branch and commit, kas now checks if the commit is
actually contained in the branch. This is especially relevant when using
lockfiles, where the final ref is locked (the commit), but still
tracking information about the branch is provided. By checking this, we
harden kas against rebaising upstream branches where a commit might
still be reachable after a rebase, but not via the specified branch.

While adding this a refactoring of the checkout implementation was
required, which also simplifies the code w.r.t the number of branches.

Tests for invalid combinations are added, both for git and mercurial.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>