tests: use monkeypatch to change cwd
Previously we used os.chdir together with a custom fixture to
change the cwd in the tests. As monkeytest already provides a pattern
for this, we switch it over.

While doing so, we introduce a new custom fixture monkeykas, which
returns a monkeypatch instance and can easily be extended with further
test preparations and cleanups.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>