tests: isolate test-env from global env
This patch fixes issues when running the tests on systems that already
have some KAS_* variables set. We implement that using monkeypatch and
an env cleanup fixture where the KAS_* variables are removed from the
environment the test is then executed in.

This also enables us to automatically clean modifications of the env
vars between the stages of the tests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>