run commands without live-output as default
This patch changes the default of the run_cmd and run_cmd_async commands
to not emit the live output. For almost all commands the output is
meaningless as the input is only emitted in debug mode. For debugging
the output can still be seen when running in debug mode.
As the build plugin is the only where the live output is actually
needed, we set the default of the liveoutput parameter to False and only
enable it there.

In the past, this output has provided valuable information, but with the
recent improvements around error handling and logging in kas, the
raw output is not needed anymore.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>