refactor: move git config part to dedicated function
This patch moves the setup of the .gitconfig file into a dedicated
function and generates the config with the GitConfigParser utility.

While this is a pure refactoring, the GitConfigParser also ensures
that arguments to the config helpers are correctly escaped in the
config file.

The GitPython is added as a mandatory dependency for kas.
This python git package is widely available across distros, small
and maintained since years.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>