issue warning on requested update of pinned repo
When pinning a repo to a specific commit (using the 'commit') config
item, this repo cannot be updated. This is different from using the
override: commit entry of the lockfiles, which is ignored on updates
(technically the lockfile is ignored).

As this behavior already confused users, file a warning when running an
update on pinned repos. When using the deprecated refspec, no warning is
issues, as we cannot distinguish commits, branches and tags.

Reported-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>