Dockerfile: Build against snapshot.debian.org
Switch the debian sources to the snapshot version that the container
image used as well. This will ensure that all packages installed later
on during the build have clearly defined versions.

We only do the apt sources pinning if a DEBIAN_TAG is provided that
contains a date. Otherwise, stick with latest version. As ARG goes out
of scope after the FROM, we need to repeat that statement.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>