Dockerfile: Address non-reproducible installation of git-lfs
The postinst script of git-lfs runs git lfs install, and that does not
always adds the lfs filter rules in the same order to /etc/gitconfig. To
ensure a stable result, do this step-wise via git config instead. A bit
ugly as the filter rules may change in future versions, but it's
currently the best we can do.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>