Dockerfile: Cache the apt cache locally
This will accelerate rebuilds, specifically when switching to
snapshot.debian.org which is known to be slow. As /var/cache/apt is now
managed outside of the container, we can and want to drop the calls of
apt-get clean and the removal of /var/lib/apt/lists.

As sharing=locked seemed to cause issues with buildah, make the mode
configurable for unofficial non-concurrent local builds via build arg
CACHE_SHARING=shared.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>