docs: move plugins to separate files
While doing so, we also enable the subcommand feature of argparse per
plugin and disable printing the subcommand options in the kas command
line section.

As this is a preparation to have dedicated manpages for the plugins.
To avoid further modifications on the same files, we also add the html
only tags for now.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>