make license detectable by licensee
The licensee tool is used by github (and others) to detect the name,
version and attribution of a repositories license. In case of github,
this data is then shown in the summary page, making it easily available
to users. As kas used a non standard header of the license, the tooling
was not able to detect the license with a sufficiently high confidence
to show this as MIT.

This patch makes the header standard and by that reaches a confidence
level of 100% for MIT in the licensee tool.

This patch does NOT change the license itself in any form. It is just
about making it accessible to tooling.

Closes #86
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>