docs: create individual manpages
For that, we add conditional sections in the docs to avoid duplications
in the manpages and the html documentation. When building the manpages,
we add the _man folder to the RST search path. By that, we avoid
duplicate entries and unreferenced files warnings in the html
documentation.

Documentation parts which are not relevant for a user of kas are
excluded as well (e.g. the devguide).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>