add bash completion for kas in docker
This patch adds a bash completion file for kas which is generated with
shtab. As the build-only dependency python3-shtab is quite small, we
just add it to the base container instead of splitting the build.

As of today, this completion file only has limited support for
sub-parsers due to a bug in shtab 1.5.5, but once 1.5.8 is accepted,
subparsers will also work out of the box.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>