fix(ssh-setup-agent): correctly parse result
In dbd32b04 we switched to run_cmd to execute the ssh-agent. However,
this wrapper returns the output as a single string instead of
line-chunks. By that, we iterated over the single characters instead of
the lines which broke the key-value parsing.

This patch fixes this by splitting the string at the end-of-line
delimiter.

Reported-by: Jasper Orschulko <Jasper.Orschulko@iris-sensing.com>
Closes #110
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>