scripts: Improve reproduction script by selecting the right repo revision
When reproducing an image which revision is not locally checked out in
git, build-container.sh was using the wrong git revision as well which
caused a false negative of the reproduction test. Address that by
checking out the git revision corresponding to the image tag that was
selected. For that purpose, build-container.sh is extended by
--git-refspec to select a specific revision in cloned repo used for the
build.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>