make file permissions on credentials more strict
The automatically generated credential files should only be readable by
the current user (instead of everyone). While this is currently not
enforced by the aws cli or git, more and more tools do not accept
credential files with incorrect permissions any longer (e.g. ssh,
wireguard).

As kas-home is created with chmod 0700, everything below it is already
safe even if the permissions inside are a bit too loose. So this
addresses no security issue.

Proposed-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>