Don't add comments to .netrc
A recent change to OE core causes bitbake to fallback to ~/.netrc when
BB_HASHSERVE_USERNAME or BB_HASHSERVE_PASSWORD are not set.

This exposes us to an error when running in GitLab because kas has been
inserting a comment into this file.

netrc.NetrcParseError: bad toplevel token 'appended'

Don't use comments in .netrc as they are not a documented feature.

Signed-off-by: Adam Johnston <adam.johnston@arm.com>