Don't add comments to .netrc
A recent change to OE core causes bitbake to fallback to netrc when
BB_HASHSERVE_USERNAME or BB_HASHSERVE_PASSWORD are not set.

This exposes us to an error when running in GitLab because kas has been
inserting a comment into ~/.netrc.

netrc.NetrcParseError: bad toplevel token 'appended'

Remove the comment from .netrc as it relies on undocumented behaviour.

Signed-off-by: Adam Johnston <adam.johnston@arm.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>