docs: do not build docs against installed version
Previously, the docs were built against an installed version of kas, as
the local path to kas was appended to the python module path. By that,
the modules first were searched in the global system installation and
only if not found the current ones were used.

This patch fixes this by prepending the current path to kas to the
module path before building the documentation. This is also proposed
by sphinx-doc.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>