document scope of environment variables
The docs already provide an overview of all environment variables
that are processed or forwarded by kas. However, it is not documented
which scopes is used or made available in.

This patch improves the documentation by stating the scope of each group
of variables. While changing this, we also align the capitalization of
the headlines in command-line.rst.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
[Jan: minor reformatting for NO_PROXY entry]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>