show steps for the --skip argument in --help
This is the most appropriate place to show the list of STEP options, as
the user is already requesting help with usage. Alternatives would be
more implicit (specifying an invalid STEP or the empty string or the
like) or would contribute to command proliferation (new list-build-steps
command).

The available choices are shown by default when using argparse, but they
are shown in the place of the METAVAR, but this gets badly formatted by
the help formatter.

If we instead add them to the help string in the usual way, by adding
%(choice)s to the help string, this causes bad formatting and
duplication in the docs, due to the way sphinx-argparse handles choice
actions.

sphinx-argparse does not use the help formatter for actions, however, so
if we implement the addition to the help string there, the docs are left
unmodified. Note that, of necessity, this formatter applies to all
subcommands, and all arguments, but only has an effect on the --skip
argument at this time.

Example output in a 72-column terminal to fit commit message guidelines:

  --skip STEP         Skip build steps. To skip more than one step,
                      use this argument multiple times. Possible
                      choices: setup_dir, setup_home, init_setup_repos,
                      repo_setup_loop, finish_setup_repos,
                      repos_checkout, repos_apply_patches,
                      setup_environ, write_bbconfig

Signed-off-by: Christopher Larson <chris.larson@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>