docs: major refactoring of document structure
The documentation did grow a lot in recent releases. This led to an
imbalance of the amount of text per file / chapter. Further, we had
too few referencable files to properly extract the content for the
manpages on a per-file basis. By that, many in-line switches (only html,
only man) needed to be introduced. Also, the documentation was a bit
clumsy to read online due to the varying amount of text per page.

This patch restructures the documentation by grouping parts into files,
which are then referenced in a TOC. Further, we use a sub TOC for the
userguide as this contains a lot of content. With that, we are able to
reference files from the userguide directly from the man pages, making
it possible to solve various issues around the sphinx include directive
(e.g. including at the wrong headline level).

The patch is documentation only and does not contain any functional
changes.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>