use private homedir in tests
Previously the tests used the HOME environment from the calling user.
For some tooling (like git), this is problematic as the tests are
executed against a properly configured git instead of a bare one.
Also other tools might access well-known paths below HOME to get parts
of their configuration.

This patch now relocates the homedir to a private (temporary) one when
executing the tests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>