repos: try to resolve url and commit of local repos
While kas already checks if the first kas project-config is in a
directory under VCS, this information is not associated with the
internal repo dictionary. The same applies to repos which are added
using 'path'.

As a preparation to precisely describe the current build environment, we
now also add this tracking information to the repo dictionary. By that,
it becomes available to tools like the dump plugin.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>