repos: add property to check if repo is dirty
For some operations, kas already checks if a repo contains changes.
However, this information was only available within the repo processing
logic itself.

As this information is relevant for other users as well (like the dump
plugin), we make it a property of the repo class. Existing users are
migrated to the new property. Repos which are not under VCS are always
considered 'dirty', as we cannot check for local changes.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>