add artifacts key to describe build artifacts
If this key is set, the output is searched for the specified artifacts.
The build is marked as failed if not for every entry at least one file
is found. For files that are found, but which have a mtime not within
the time duration of the build, a warning is issued.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>