kas-container: Handle missing extra argument in subcommands gracefully
When specifying a subcommand option that takes an extra argument without
that, kas-container currently just terminates without any error. That is
because "shift 2" just bails out dues to "set -e". Handle this case by
shifting the option first and then the extra argument. When the latter
fails, we simply make the subcommand dump its own help.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>